+++ service-base/src/main/java/org/eclipse/hono/service/resourcelimits/ResourceLimitChecks.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package org.eclipse.hono.service.plan;
+package org.eclipse.hono.service.resourcelimits;
 
 import org.eclipse.hono.client.ServiceInvocationException;
 import org.eclipse.hono.util.TenantObject;
